+++ tests/common/features/schema.rs
@@ -63,6 +63,7 @@ pub async fn create_tables(db: &DatabaseConnection) -> Result<(), DbErr> {
         create_categories_table(db).await?;
         #[cfg(feature = "postgres-vector")]
         create_embedding_table(db).await?;
+        create_host_network_table(db).await?;
     }
 
     Ok(())
@@ -473,6 +474,36 @@ pub async fn create_collection_table(db: &DbConn) -> Result<ExecResult, DbErr> {
     create_table(db, &stmt, Collection).await
 }
 
+pub async fn create_host_network_table(db: &DbConn) -> Result<ExecResult, DbErr> {
+    let stmt = sea_query::Table::create()
+        .table(host_network::Entity)
+        .col(
+            ColumnDef::new(host_network::Column::Id)
+                .integer()
+                .not_null()
+                .auto_increment()
+                .primary_key(),
+        )
+        .col(
+            ColumnDef::new(host_network::Column::Hostname)
+                .string()
+                .not_null(),
+        )
+        .col(
+            ColumnDef::new(host_network::Column::Ipaddress)
+                .inet()
+                .not_null(),
+        )
+        .col(
+            ColumnDef::new(host_network::Column::Network)
+                .cidr()
+                .not_null(),
+        )
+        .to_owned();
+
+    create_table(db, &stmt, HostNetwork).await
+}
+
 pub async fn create_pi_table(db: &DbConn) -> Result<ExecResult, DbErr> {
     let stmt = sea_query::Table::create()
         .table(pi::Entity)
@@ -745,10 +776,21 @@ pub async fn create_value_type_table(db: &DbConn) -> Result<ExecResult, DbErr> {
                 .integer()
                 .not_null(),
         )
+        .col(
+            ColumnDef::new(value_type::value_type_general::Column::Tag1)
+                .string()
+                .not_null(),
+        )
+        .col(
+            ColumnDef::new(value_type::value_type_general::Column::Tag2)
+                .string()
+                .not_null(),
+        )
         .to_owned();
 
     create_table(db, &general_stmt, value_type::value_type_general::Entity).await
 }
+
 pub async fn create_value_type_postgres_table(db: &DbConn) -> Result<ExecResult, DbErr> {
     let postgres_stmt = sea_query::Table::create()
         .table(value_type::value_type_pg::Entity)
@@ -765,7 +807,7 @@ pub async fn create_value_type_postgres_table(db: &DbConn) -> Result<ExecResult,
                 .not_null(),
         )
         .col(
-            ColumnDef::new(json_vec::Column::StrVec)
+            ColumnDef::new(value_type::value_type_pg::Column::StrVec)
                 .array(sea_query::ColumnType::String(StringLen::None))
                 .not_null(),
         )